ParamAccessor now replaces all(?) params in maj7, a lot of space savings and i think speed increase. also some build warnings gone, some bugs/nits fixed, for example: - wrong return type in ParamAccessor::GetBoolValue - bad cast when subclassing window in exe music main - optimizations were i think not working 100% correctly for LFO. - possibly some values are now being clamped as they should

+++ WaveSabreExecutableMusicPlayer/renderer.hpp
@@ -67,7 +67,7 @@ struct Renderer
         WaveHDR =
         {
           (LPSTR)gpBuffer,
-          bufferSizeBytes,
+          (DWORD)bufferSizeBytes,
           0,
         };
     }
